Feature20

+++ classes/User.php
@@ -635,20 +635,29 @@ header('Location: home.php');
         //remove topics from user in session that are empty
         
         //remove boards from user in session
+        $statement4 = $conn->prepare("DELETE FROM boards WHERE user_id = :userid;");
+        $statement4->bindValue(":userid", $_SESSION['userid']);
+        $statement4->execute();
         
         //remove follow relations from user in session
+        $statement5 = $conn->prepare("DELETE FROM follows WHERE follower = :userid;");
+        $statement5->bindValue(":userid", $_SESSION['userid']);
+        $statement5->execute();
         
         //remove like relations from user in session
+        $statement6 = $conn->prepare("DELETE FROM likes WHERE UserID = :userid;");
+        $statement6->bindValue(":userid", $_SESSION['userid']);
+        $statement6->execute();
         
         //unlink user picture
-        /*if ($_SESSION["image"] != "profile_placeholder.png") {
+        if ($_SESSION["image"] != "profile_placeholder.png") {
             unlink("../images/uploads/userImages/" . $_SESSION["image"]);
         }
         
         //remove user in session
         $statement7 = $conn->prepare("DELETE FROM users WHERE email = :email;");
         $statement7->bindValue(":email", $_SESSION['user']);
-        $statement7->execute();*/
+        $statement7->execute();
         
         session_unset();
